==> 2. Topic extraction.ipynb <==
import pandas as pd

papers = pd.read_csv('papers.csv').iloc[:, 1:]
patents = pd.read_csv('patents.csv').iloc[:, 1:]

display(papers.head())
patents.head()

## Subset

# import numpy as np

# data = pd.read_csv('papers.csv', nrows=100)

# def text_process(data, columns):
#     d = data[columns].apply(lambda s: s.str.lower().str.replace('[\W]', ' ', regex=True))
#     d = d.fillna('').apply(lambda s: ' '.join(s.tolist()), axis=1)
#     return d
#     return d.apply(lambda t: ' ' + t + ' ').str.replace('\s+', ' ', regex=True)

# def digital_topic_extraction(text, topics):
#     topics = '( ' + ' | '.join(topics) + ' )'
#     d = text.str.findall(topics).apply(lambda l: set(l))
#     d = d.apply(lambda l: '/'.join(l))
#     return pd.DataFrame({'topics': d})

# def data_process(data, columns, topics, drop=False):
#     d = text_process(data, columns)
#     d = digital_topic_extraction(d, topics)
#     d = d.merge(data, left_index=True, right_index=True)
#     if drop:
#         return d.drop(columns, axis=1)
#     else:
#         return d
    
# text_process(data, ['title', 'abstract', 'author_keywords', 'keywords_plus', 'wos_categ'])[11]

import numpy as np

data = papers.iloc[:1000, :].copy()

def text_process(data, columns):
    d = data[columns].apply(lambda s: s.str.lower().str.replace('[\W]', ' ', regex=True))
    d = d.fillna('').apply(lambda s: ' '.join(s.tolist()), axis=1)
    return d
    return d.apply(lambda t: ' ' + t + ' ').str.replace('\s+', ' ', regex=True)

def digital_topic_extraction(text, topics):
    topics = '( ' + ' | '.join(topics) + ' )'
    d = text.str.findall(topics).apply(lambda l: set(l))
    d = d.apply(lambda l: '/'.join(l))
    return pd.DataFrame({'topics': d})

def data_process(data, columns, topics, drop=False):
    d = text_process(data, columns)
    d = digital_topic_extraction(d, topics)
    d = d.merge(data, left_index=True, right_index=True)
    if drop:
        return d.drop(columns, axis=1)
    else:
        return d
    
data_process(data, columns=['title', 'abstract', 'author_keywords', 'keywords_plus', 'wos_categ'], 
             topics=['analysis', 'science', 'hydrology'], drop=True)

## Whole data

import json
import copy
import collections
import re

with open('topics.json', 'r') as f:
    topics = json.load(f)
    
t = copy.deepcopy(topics)

for k in t:
    for kk in t[k]:
        topics[k][kk] = re.sub("-", "", t[k][kk]).lower().split(', ')
        assert len(topics[k][kk]) == len(set(topics[k][kk])), "Duplicates inside a field"
        
digital_topics = []
for k in t:
    for kk in t[k]:
        digital_topics += t[k][kk].split(', ')
dups = [item for item, count in collections.Counter(digital_topics).items() if count > 1]
print(len(dups), "Duplicates across fields:\n", dups)

topics

papers2 = data_process(papers, columns=['title', 'abstract', 'author_keywords', 'keywords_plus', 'wos_categ'], 
                       topics=digital_topics, drop=True)

papers2 = papers2[papers2['topics'] != '']
papers2.to_csv('papers_topics.csv')

patents2 = data_process(patents, columns=['title', 'abstract', 'keywords', 'categories'],
                        topics=digital_topics, drop=True)
patents2 = patents2[patents2['topics'] != '']
patents2.to_csv('patents_topics.csv')
