==> 3. Data prep.ipynb <==
import pandas as pd

papers = pd.read_csv('data/papers_topics.csv').iloc[:, 1:]
patents = pd.read_csv('data/patents_topics.csv').iloc[:, 1:]

## Subset - Unique topics

import re

data = papers.sample(5000).copy()

def unique_topics(series):
    t = series.str.strip().str.split(' / ').tolist()
    t = set([x for y in t for x in y])
    print(len(t), 'topics')
    return t

unique_topics(data['topics'])

**Remove plurals**

def remove_plurals(series):
    return series.str.replace('( \w+)s ', '\g<1> ', regex=True).str.replace(' (\w+)s $', ' \g<1> ', regex=True)

data['topics'] = remove_plurals(data['topics'])
unique_topics(data['topics'])

**Group related topics**

def map_topics(series, mapper):
    series = series.str.title()
    for topic, regex in mapper.items():
        series = series.str.replace(regex, topic, regex=True, flags=re.IGNORECASE)
    return series

mapper = {
    "Analysis": "analysi", " Loss ": ' los? ', "Extraction": "extrac\w+", " Modeling ": " model\w* ",
    " SaaS ": " saas? ", " IaaS ": " iaas? ", " PaaS ": " paas? ",
    "Robots": "robots?", "Robotics": "robotics?|robotsics?",
    "Classification": "classif\w+",
    "Redis ": "redi ",
    "3D Print": "3d print\w*",
    "Big": "massive|large scale",
    " AI ": " ai | artificial intellig\w+ | Machine Intelligence | Intelligent Machine ",
    " AUV ": " auv | autonomous underwater vehicle | underwater vehicle | Unmanned Underwater ",
    "Back-propagation": "back propagation|backpropagation",
    " Clouds ": " cloud \w+ | Multi Cloud ",
    "Cluster Computing ": "cluster \w+ ",
    " DBM ": " dbm ",
    " Deep Learning ": " deep learning | dl ",
    " Generative Adversarial Network ": " gans? | generative adversarial networks? | generative adversarial ",
    " GPU ": " gpu | graphics? processing units? ",
    "Hardware Acceleration": "hardware acceler\w+",
    " Machine Learning ": " ml | machine learning ",
    "Markov ": "markov \w+ ",
    " Multi-layer perceptron ": " multi layer perceptrons? | mlp | multilayer perceptron ",
    " NLP ": " natural language processing | nlp ",
    "Nearest Neighbors": "nearest neighbou?rs?",
    "Object Recognition": "object detection|object recognition|object identification",
    "Dimensionality Reduction": 'factor analysis|principal components?',
    "QPU": "quantum processing units?",
    " SVM ": " support vector | svm ",
    " UAM ": " drones? | uav | unmanned aerial vehicle | unmanned aircraft\s?\w* | uas | unmanned air vehicle ",
    " AWS ": " aw ",
    " Optimization Algorithm ": " adam | gradient \w{1,2}scent ",
    "Boosting": "Boosting Algorithm|Gradient Boosting",
    " Convolutional Networks ": " cnns? | Convolutional Layers? | Convolutional Neural Net\w* ",
    "Data Center": "data centre",
    "Deep Learning": "Deep Neural Network|dnn",
    "Text Embedding": "doc2vec|word2vec|seq2seq",
    " RNN ": " gru | lstm | rnn | recurrent neural net\w* | Gated Recurrent Unit | Long Short Term Memory ",
    " UGV ": " ugv | Unmanned Ground Vehicle | Autonomou Vehicle | Self Driving | Autonomou car ",
    " Voice Recognition ": " Voice Activity Detection | text2speech | text to speech | Speech Recognition ",
    " Search Algorithm ": " Tree Search | search trees? ",
#     "Computer Vision": "text\s?to\s?image|Image Classification|Image Recognition|Image Segmentation|Gesture Recognition|",
    " Trees ": " regression trees? | classification trees? | random forests? ",
    "Transformers & Attention": "Attention Mechanisms?|Attention layers?|transformer architectures?|transformer net\\w+",
    "Supercomputers": "Super\s?comput\w+",
    "Quantum Computing": "Quantum Comput\w+",
    "Neuromorphic Computing": "Neuromorphic Comput\w+",
    " Reinforcement Learning ": " rl | q learning | policy gradient | policy optimization ",
    " BD & Cloud Solutions ": " On Demand Computing | hadoop | spark | mysql | cassandra | neo[45]j | mongodb | hive | kafka | ec2 | elasticsearch | bigtable | hbase | oracle | microsoft sql | redis | ibm db2 | microsoft azure | google cloud | sqlite | mariadb | splunk | dynamodb ",
    " NRDBMS & NoSQL ": " nosql | nrdbm?s? | non\s?Relational Database ",
    " RDBMS & SQL ": " rdbm?s? | Relational Database | sql | mysql | sqlite ",
    " Database ": " dbm ",
    "Cyber Infrastructure": "Cyber\s?Infrastructure"
}

data['topics'] = map_topics(data['topics'], mapper)
unique_topics(data['topics'])

data.head()

## Whole data

**Science**

papers2 = papers.copy()
unique_topics(papers2['topics'])

papers2['topics'] = remove_plurals(papers2['topics'])
unique_topics(papers2['topics'])

papers2['topics'] = map_topics(papers2['topics'], mapper)
unique_topics(papers2['topics'])

papers2.to_csv('data/papers_topics2.csv', index=False)
papers2.head()

**Industry**

patents2 = patents.copy()
unique_topics(patents2['topics'])

patents2['topics'] = remove_plurals(patents2['topics'])
unique_topics(patents2['topics'])

patents2['topics'] = map_topics(patents2['topics'], mapper)
unique_topics(patents2['topics'])

patents2.to_csv('data/patents_topics2.csv', index=False)
patents2.head()

## Data preparation

import pandas as pd

papers = pd.read_csv('data/papers_topics2.csv')
patents = pd.read_csv('data/patents_topics2.csv')

papers = papers[~(papers['research_areas'].isna() | papers.AU.isna())]

import numpy as np
import re

def count_topics(data, science=False):
    series = data["topics"]
    t = series.str.strip().str.split(' / ').tolist()
    t = pd.Series([x for y in t for x in y])
    vc, vcp = t.value_counts(), t.value_counts(normalize=True)
    d = pd.DataFrame({'count': vc, 'proportion': vcp})
    if science:
        d['citations'] = pd.Series(vc.index).\
                            apply(lambda key: data.loc[data["topics"].str.contains(key), "n_citations"].sum(skipna=True)).\
                            values
    return d

def strata_count(data, strata='year', datat='patents', string=False, rank=None):
    if string:
        col = strata
        strata = data[strata].str.split('; ').tolist()
        strata = set([x for y in strata for x in y])
        for stratum in strata:
            if stratum in rank.index: 
                ct = count_topics(data.loc[data[col].str.contains(stratum, regex=True)]) if datat=='patents' else \
                     count_topics(data.loc[data[col].str.contains(stratum, regex=True)], science=True)
                path = f"data/counts/{datat}_count_{col}_{rank[stratum] if rank is not None else 'a'}_" + \
                        re.sub('\W', '_', stratum) + ".csv"
                ct.to_csv(path, index_label="topic")
    else:
        for stratum in data[strata].unique():
            ct = count_topics(data.loc[data[strata] == stratum])if datat=='patents' else \
                 count_topics(data.loc[data[strata] == stratum], science=True)
            ct.to_csv(f"data/counts/{datat}_count_{strata}_{stratum}.csv", index_label="topic")

**Overall count**

count_topics(patents).to_csv('data/counts/patents_count.csv', index_label="topic")
count_topics(papers, science=True).to_csv('data/counts/papers_count.csv', index_label="topic")

**Count per year**

strata_count(patents, strata='year', datat='patents')
strata_count(papers, strata='year', datat='papers')

**Count per author**

patents.inventor = patents.inventor.str.split(", ', ").str.join("; ").str.replace("[',]", "", regex=True)
au = patents.inventor.str.split("; ").tolist()
au = pd.Series([x for y in au for x in y]).value_counts()[:50]
au = pd.Series(list(range(1, 51)), index=au.index)
pattern = '|'.join(au.index)

strata_count(patents[patents.inventor.str.contains(pattern, regex=True)], 
             strata='inventor', datat='patents', string=True, rank = au)

papers.AU = papers.AU.str.strip().str.replace(',', '',regex=True)
au = papers.AU.str.split('; ').tolist()
au = pd.Series([x for y in au for x in y]).value_counts()[:50]
au = pd.Series(list(range(1, 51)), index=au.index)
pattern = '|'.join(au.index)

strata_count(papers[papers.AU.str.contains(pattern, regex=True)], 
             strata='AU', datat='papers', string=True, rank = au)

**Count per Publisher (papers only)**

papers.publisher = papers.publisher.str.strip().str.replace('\W', ' ',regex=True)
pub = papers.publisher.str.split('; ').tolist()
pub = pd.Series([x for y in pub for x in y]).value_counts()[:50]
pub = pd.Series(list(range(1, 51)), index=pub.index)
pattern = '|'.join(pub.index)

strata_count(papers[papers.publisher.str.contains(pattern, regex=True)], 
             strata='publisher', datat='papers', string=True, rank = pub)

**Count per research area (papers only)**

def ra_count(data, rank, y):
    strata = data['research_areas'].str.split('\s{2,3}').tolist()
    strata = set([x for y in strata for x in y])
    ds = []
    for stratum in strata:
        d = data[data["research_areas"].str.contains(stratum)]
        t = d['topics'].str.strip().str.split(' / ').tolist()
        t = pd.Series([x for y in t for x in y])
        vc = t.value_counts()
        d = pd.DataFrame({'count': vc, 'ra': stratum, 'year': y})
        d['citations'] = pd.Series(vc.index).\
                            apply(lambda key: data.loc[data["topics"].str.contains(key), "n_citations"].sum(skipna=True)).\
                            values
        ds.append(d.reset_index())
    try:
        return pd.concat(ds)
    except:
        pass

papers.research_areas = papers.research_areas.str.strip().str.replace('\W', ' ',regex=True)
ra = papers.research_areas.str.split('\s{2,3}').tolist()
ra = pd.Series([x for y in ra for x in y]).value_counts()[:30]
ra = pd.Series(list(range(1, 31)), index=ra.index)
pattern = '|'.join(ra.index)

ds = []
for y in range(2000, 2022):
    d = ra_count(papers[(papers.year==y) & (papers.research_areas.str.contains(pattern, regex=True))], ra, y)
    ds.append(d)
res = pd.concat(ds).reset_index(drop=True)
res.columns = ['Topic', 'count', 'ScienceT', 'year', 'citations']
res.to_csv('data/counts/papers_RA.csv', index=False)
res

## Co-occurence matrix

import pandas as pd

papers = pd.read_csv('data/papers_topics2.csv')
patents = pd.read_csv('data/patents_topics2.csv')

import numpy as np
import itertools
from scipy.sparse import csr_matrix

def co_matrix(unique_topics, topics_list): 
    word_to_id = dict(zip(unique_topics, range(len(unique_topics))))
    documents_as_ids = [np.sort([word_to_id[w] for w in doc if w in word_to_id]).astype('uint32') for doc in topics_list]
    row_ind, col_ind = zip(*itertools.chain(*[[(i, w) for w in doc] for i, doc in enumerate(documents_as_ids)]))
    
    data = np.ones(len(row_ind), dtype='uint32')  
    max_word_id = max(itertools.chain(*documents_as_ids)) + 1
    
    docs_words_matrix = csr_matrix((data, (row_ind, col_ind)), shape=(len(documents_as_ids), max_word_id))  
    words_cooc_matrix = docs_words_matrix.T * docs_words_matrix 
    words_cooc_matrix.setdiag(0)
    
    return pd.DataFrame(words_cooc_matrix.todense(), index=unique_topics, columns=unique_topics)

**Industry**

topics_list = patents.topics.str.split("/").tolist()
unique_topics = set([x for y in topics_list for x in y])

cm = co_matrix(unique_topics, topics_list)
cm.to_csv("data/counts/patents_co_occ.csv", index_label="Topic")
cm.head()

for y in patents.year.unique():
    topics_list = patents.loc[patents.year == y, "topics"].str.split("/").tolist()
    unique_topics = set([x for y in topics_list for x in y])

    cm = co_matrix(unique_topics, topics_list)
    cm.to_csv(f"data/counts/patents_co_occ_{y}.csv", index_label="Topic")

**Science**

topics_list = papers.topics.str.split("/").tolist()
unique_topics = set([x for y in topics_list for x in y])

cm = co_matrix(unique_topics, topics_list)
cm.to_csv("data/counts/papers_co_occ.csv", index_label="Topic")
cm.head()

for y in papers.year.unique():
    topics_list = papers.loc[papers.year == y, "topics"].str.split("/").tolist()
    unique_topics = set([x for y in topics_list for x in y])

    cm = co_matrix(unique_topics, topics_list)
    cm.to_csv(f"data/counts/papers_co_occ_{y}.csv", index_label="Topic")
